Move component to common
