+++ src/components/App.jsx
@@ -6,7 +6,7 @@ import { connect } from 'react-redux'
 import * as actions from '../actions/appActions'
 import _ from 'lodash'
 import SearchImage from './SearchImage/SearchImageTop'
-import Circle from './SearchImage/Circle'
+import Circle from './Circle'
 import Header from './SearchImage/Header'
 import { CONTENT_CLASSES, CHANNEL_IMAGES } from '../const'
 import lang from '../lang.json'
